Refactor schema.org to have a json_ld parse module
